+++ aks/main.tf
@@ -1,7 +1,17 @@
-data "azurerm_resource_group" "node" {
+# This null_resource is required otherwise Terraform would try to read the resource group data and/or the storage 
+# account even if they were not created yet. 
+resource "null_resource" "dependencies" {
+  triggers = var.dependency_ids
+}
+
+data "azurerm_resource_group" "node_resource_group" {
   count = local.use_managed_identity ? 1 : 0
 
   name = var.logs_storage.managed_identity_node_rg_name
+
+  depends_on = [
+    resource.null_resource.dependencies
+  ]
 }
 
 data "azurerm_storage_container" "container" {
@@ -9,17 +19,21 @@ data "azurerm_storage_container" "container" {
 
   name                 = var.logs_storage.container
   storage_account_name = var.logs_storage.storage_account
+
+  depends_on = [
+    resource.null_resource.dependencies
+  ]
 }
 
 resource "azurerm_user_assigned_identity" "loki" {
   count = local.use_managed_identity ? 1 : 0
 
-  resource_group_name = data.azurerm_resource_group.node[0].name
-  location            = data.azurerm_resource_group.node[0].location
   name                = "loki"
+  resource_group_name = data.azurerm_resource_group.node_resource_group[0].name
+  location            = data.azurerm_resource_group.node_resource_group[0].location
 }
 
-resource "azurerm_role_assignment" "contributor" {
+resource "azurerm_role_assignment" "storage_contributor" {
   count = local.use_managed_identity ? 1 : 0
 
   scope                = data.azurerm_storage_container.container[0].resource_manager_id
@@ -31,25 +45,25 @@ resource "azurerm_federated_identity_credential" "loki" {
   count = local.use_managed_identity ? 1 : 0
 
   name                = "loki"
-  resource_group_name = data.azurerm_resource_group.node[0].name
+  resource_group_name = data.azurerm_resource_group.node_resource_group[0].name
   audience            = ["api://AzureADTokenExchange"]
   issuer              = var.logs_storage.managed_identity_oidc_issuer_url
   parent_id           = azurerm_user_assigned_identity.loki[0].id
-  subject             = "system:serviceaccount:${var.namespace}:loki" # "loki" is the fullnameOverride value
+  subject             = "system:serviceaccount:loki-stack:loki" # "loki" is the fullnameOverride value
 }
 
 module "loki-stack" {
   source = "../"
 
-  argocd_namespace = var.argocd_namespace
-  target_revision  = var.target_revision
-  namespace        = var.namespace
-  app_autosync     = var.app_autosync
-  dependency_ids   = var.dependency_ids
+  argocd_project      = var.argocd_project
+  argocd_labels       = var.argocd_labels
+  destination_cluster = var.destination_cluster
+  target_revision     = var.target_revision
+  app_autosync        = var.app_autosync
+  dependency_ids      = var.dependency_ids
 
-  retention       = var.retention
-  ingress         = var.ingress
-  enable_filebeat = var.enable_filebeat
+  retention = var.retention
+  ingress   = var.ingress
 
   helm_values = concat(local.helm_values, var.helm_values)
 }