+++ main.tf
@@ -13,21 +13,20 @@ resource "htpasswd_password" "loki_password_hash" {
 }
 
 resource "argocd_project" "this" {
+  count = var.argocd_project == null ? 1 : 0
+
   metadata {
-    name      = "loki-stack"
-    namespace = var.argocd_namespace
-    annotations = {
-      "devops-stack.io/argocd_namespace" = var.argocd_namespace
-    }
+    name      = var.destination_cluster != "in-cluster" ? "loki-stack-${var.destination_cluster}" : "loki-stack"
+    namespace = "argocd"
   }
 
   spec {
-    description  = "Loki application project"
+    description  = "Loki application project for cluster ${var.destination_cluster}"
     source_repos = ["https://github.com/camptocamp/devops-stack-module-loki-stack.git"]
 
     destination {
-      name      = "in-cluster"
-      namespace = var.namespace
+      name      = var.destination_cluster
+      namespace = "loki-stack"
     }
 
     orphaned_resources {
@@ -47,8 +46,12 @@ data "utils_deep_merge_yaml" "values" {
 
 resource "argocd_application" "this" {
   metadata {
-    name      = "loki-stack"
-    namespace = var.argocd_namespace
+    name      = var.destination_cluster != "in-cluster" ? "loki-stack-${var.destination_cluster}" : "loki-stack"
+    namespace = "argocd"
+    labels = merge({
+      "application" = "loki-stack"
+      "cluster"     = var.destination_cluster
+    }, var.argocd_labels)
   }
 
   timeouts {
@@ -59,20 +62,21 @@ resource "argocd_application" "this" {
   wait = var.app_autosync == { "allow_empty" = tobool(null), "prune" = tobool(null), "self_heal" = tobool(null) } ? false : true
 
   spec {
-    project = argocd_project.this.metadata.0.name
+    project = var.argocd_project == null ? argocd_project.this[0].metadata.0.name : var.argocd_project
 
     source {
       repo_url        = "https://github.com/camptocamp/devops-stack-module-loki-stack.git"
       path            = "charts/loki-microservice"
       target_revision = var.target_revision
       helm {
-        values = data.utils_deep_merge_yaml.values.output
+        release_name = "loki"
+        values       = data.utils_deep_merge_yaml.values.output
       }
     }
 
     destination {
-      name      = "in-cluster"
-      namespace = var.namespace
+      name      = var.destination_cluster
+      namespace = "loki-stack"
     }
 
     sync_policy {