+++ variables.tf
@@ -2,22 +2,28 @@
 ## Standard variables
 #######################
 
-variable "argocd_namespace" {
-  description = "Namespace used by Argo CD where the Application and AppProject resources should be created."
+variable "argocd_project" {
+  description = "Name of the Argo CD AppProject where the Application should be created. If not set, the Application will be created in a new AppProject only for this Application."
   type        = string
-  default     = "argocd"
+  default     = null
+}
+
+variable "argocd_labels" {
+  description = "Labels to attach to the Argo CD Application resource."
+  type        = map(string)
+  default     = {}
+}
+
+variable "destination_cluster" {
+  description = "Destination cluster where the application should be deployed."
+  type        = string
+  default     = "in-cluster"
 }
 
 variable "target_revision" {
   description = "Override of target revision of the application chart."
   type        = string
-  default     = "v5.0.0" # x-release-please-version
-}
-
-variable "namespace" {
-  description = "Namespace where the applications's Kubernetes resources should be created. Namespace will be created in case it doesn't exist."
-  type        = string
-  default     = "loki-stack"
+  default     = "v7.0.1" # x-release-please-version
 }
 
 variable "helm_values" {
@@ -58,12 +64,6 @@ variable "ingress" {
     allowed_ips    = optional(list(string), [])
   })
   default = null
-}
-
-variable "enable_filebeat" {
-  description = "" # TODO
-  type        = bool
-  default     = false
 }
 
 variable "retention" {